requirements.txt
open-interpreter
